Added default constructor
Default constructor creates an 'empty' list. An empty list is currently defined only in terms of it's size. I.e., a list is empty if and only if it's size is zero. Note, a list that is constructed empty will have an uninitialized m_head->value until it is added to, while a list that becomes empty will have whatever value was in m_head to start with.